Auto login after registration
This patch also introduces a new login mode UserGroupsUser::login('auto').
This may be useful in other use cases, such as authentication via oAuth.

Independent modules may perform any kind of authentication and, after
successful verification of the user identity, invoke $user->login('auto')
to let userGroups module perform the necessary steps for signin in.

The varialbe $user must be an instance of UserGroupsUser.
The fields username and password are compared against the values in the
database (password is also compared as it is, i.e. not hashed).